Move MessageConsumer to clients
